backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
celery[redis]==5.4.0
redis==5.2.0
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
python-multipart==0.0.12
websockets==13.0
requests==2.32.0
librosa==0.10.2
matplotlib==3.9.0
pillow==10.4.0
numpy==1.26.4
python-dotenv==1.0.1
aiofiles==24.1.0
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
bcrypt==4.2.0
httptools==0.6.1
